placing google relate code in the same file
